requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
packaging==23.2
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
dj-database-url>=2.0,<3.0
whitenoise[brotli]>=6.0,<7.0
psycopg[c]; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"